rename the previous env_params.py to env_params_old.py
