+++ echopype/tests/echodata/test_echodata.py
@@ -9,7 +9,7 @@ from datatree import DataTree
 from zarr.errors import GroupNotFoundError
 
 import echopype
-from echopype.calibrate.env_params import EnvParams
+from echopype.calibrate.env_params_old import EnvParams
 from echopype.echodata import EchoData
 from echopype import open_converted
